Update InlineLinkExtractor.php
Todo's added for output formatting.
Todo added for replacement of deprecated function.

Code modified for easy creation of a comma separated list. But perhaps better to return an array so the downstream function can determine what to do with the raw data. Note that right-to-left languages expects a different order or the text.